Move profile and theme sections to separate activities
