+++ app/src/main/java/com/example/app_s9/PreferencesActivity.kt
@@ -1,0 +1,49 @@
+package com.example.app_s9
+
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatDelegate
+import androidx.appcompat.widget.SwitchCompat
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+
+class PreferencesActivity : AppCompatActivity() {
+
+    private lateinit var sharedPreferencesHelper: SharedPreferencesHelper
+    private lateinit var viewModel: MainViewModel
+    private lateinit var switchDarkMode: SwitchCompat
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        sharedPreferencesHelper = SharedPreferencesHelper(this)
+        val themeMode = sharedPreferencesHelper.getInt(
+            SharedPreferencesHelper.KEY_THEME_MODE,
+            AppCompatDelegate.MODE_NIGHT_NO
+        )
+        AppCompatDelegate.setDefaultNightMode(themeMode)
+
+        enableEdgeToEdge()
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_preferences)
+
+        viewModel = ViewModelProvider(this, object : ViewModelProvider.Factory {
+            override fun <T : ViewModel> create(modelClass: Class<T>): T {
+                @Suppress("UNCHECKED_CAST")
+                return MainViewModel(sharedPreferencesHelper) as T
+            }
+        })[MainViewModel::class.java]
+
+        switchDarkMode = findViewById(R.id.switchDarkMode)
+        switchDarkMode.isChecked = viewModel.getThemeMode() == AppCompatDelegate.MODE_NIGHT_YES
+
+        switchDarkMode.setOnCheckedChangeListener { _, isChecked ->
+            val mode = if (isChecked) {
+                AppCompatDelegate.MODE_NIGHT_YES
+            } else {
+                AppCompatDelegate.MODE_NIGHT_NO
+            }
+            viewModel.saveThemeMode(mode)
+            AppCompatDelegate.setDefaultNightMode(mode)
+        }
+    }
+}
