+++ app/src/main/res/menu/main_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/action_profile"
+        android:title="Perfil" />
+    <item
+        android:id="@+id/action_preferences"
+        android:title="Preferencias" />
+</menu>
